# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BLACKSENTINEL AI Platform API",
    "description": "Cyber Cognitive Intelligence Engine - Enterprise Security AI Platform",
    "version": "1.0.0",
    "contact": {
      "name": "BlackSentinel AI Team",
      "email": "api@blacksentinel.ai"
    }
  },
  "servers": [
    {
      "url": "https://ai.blacksentinel.ai/api/v1",
      "description": "Production"
    },
    {
      "url": "https://staging-ai.blacksentinel.ai/api/v1",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/memory/store": {
      "post": {
        "tags": ["Memory Service"],
        "summary": "Store a memory entry",
        "operationId": "storeMemory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryStoreRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Memory stored"
          }
        }
      }
    },
    "/memory/retrieve": {
      "post": {
        "tags": ["Memory Service"],
        "summary": "Retrieve relevant memories",
        "operationId": "retrieveMemory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryRetrieveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retrieved memories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryRetrieveResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MemoryStoreRequest": {
        "type": "object",
        "required": ["content", "type", "tenantId"],
        "properties": {
          "content": { "type": "string" },
          "type": { "type": "string", "enum": ["immediate", "operational", "historical"] },
          "tenantId": { "type": "string" },
          "metadata": { "type": "object" }
        }
      },
      "MemoryRetrieveRequest": {
        "type": "object",
        "required": ["query", "tenantId"],
        "properties": {
          "query": { "type": "string" },
          "tenantId": { "type": "string" },
          "limit": { "type": "integer", "default": 10 }
        }
      },
      "MemoryRetrieveResponse": {
        "type": "object",
        "properties": {
          "entries": { "type": "array" },
          "totalFound": { "type": "integer" }
        }
      },
...
    }
  }
}